+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This file is used to list changes made in each version of the percona cookbook.
 
 ## Unreleased
 
+- Fix adjustment of open files limit within systemd unit
+
 ## 3.4.0 - *2023-06-07*
 
 - Avoid changing the perms on /tmp